+++ include/types/limits.hpp
@@ -2,7 +2,7 @@
 #define ATG_PALACE_ENGINE_LIMITS_HPP
 
 namespace palace {
-static constexpr size_t size_t_max = -1;
+static constexpr size_t size_t_max = static_cast<size_t>(-1);
 }
 
 #endif// ATG_PALACE_ENGINE_LIMITS_HPP
