+++ test/sanity_test.cpp
@@ -1,3 +1,3 @@
 #include <gtest/gtest.h>
 
-TEST(SanityTest, SanityTest) { EXPECT_EQ((float) 1 + 1.0, 2); }
+TEST(SanityTest, SanityTest) { EXPECT_EQ(1 + 1, 2); }
